add collectd network plugin
In order to see network metrics, we need to install the collectd plugin. Technically,
we could install this en lieu of collectd, as I am pretty sure it would pull collectd
as a dependency. But in my opinion, installing it after collectd makes it clearer what
is going on, even though it is slightly less efficient.

Signed-off-by: Glauber Costa <glommer@scylladb.com>