specify rate instead of threads
Currently we specify a threads variable, and let the user specify the
number of threads to run. However, in cassandra-stress the threads
parameters is part of the "rate" option, that has other options as well.

One way to handle that would be to keep the threads variable and add a
second optional variable with extraneous options. However, in line with
other tunables that we already have (like command_options and
stress_options), I opted for just exposing the whole -rate as a unit.

Signed-off-by: Glauber Costa <glauber@scylladb.com>